requirements.txt
numpy
PySide6
pytest
absl-py
pyyaml
msgpack
